workflow -> mapreduce

+++ cloudsimex-mapreduce/src/org/cloudbus/cloudsim/mapreduce/MapReduceDatacenter.java
@@ -41,7 +41,7 @@ public class MapReduceDatacenter extends Datacenter {
 	Hashtable<Integer,Vm> vmTable;
 	Hashtable<Long,Channel> vmChannelTable;
 	Hashtable<Vm,Long> vmCreationTime;
-	Hashtable<Vm,Long> vmPrice;
+	Hashtable<Vm,Double> vmPrice;
 	
 	Random random;
 	Random seedGenerator;
@@ -61,7 +61,7 @@ public class MapReduceDatacenter extends Datacenter {
 		this.basicCpuUnit = basicCpuUnit;
 		this.budget = 0;
 		this.vmCreationTime = new Hashtable<Vm,Long>();
-		this.vmPrice = new Hashtable<Vm,Long>();
+		this.vmPrice = new Hashtable<Vm,Double>();
 				
 		this.random = new Random(seed);
 		this.seedGenerator = new Random(seed);
@@ -125,7 +125,7 @@ public class MapReduceDatacenter extends Datacenter {
 		//do last accounting for the vm
 		Vm vm = (Vm) ev.getData();
 		long startTime = vmCreationTime.remove(vm);
-		long price = vmPrice.remove(vm);
+		Double price = vmPrice.remove(vm);
 		long useInHours = updateVmUsage(startTime,price);
 		Log.printLine(CloudSim.clock()+": VM #"+vm.getId()+" destroyed. Total usage time: "+useInHours+" h. Cost = $"+price+" cents/hour.");
 		//vmTable.remove(vm.getId());
@@ -275,11 +275,11 @@ public class MapReduceDatacenter extends Datacenter {
 		return vmChannelTable.get(key);
 	}
 	
-	private long getPrice(Vm vm) {
-		Hashtable<Vm,Integer> vmOffersTable = vmOffers.getVmOffers();
-		int cost=0;
+	private Double getPrice(Vm vm) {
+		Hashtable<Vm,Double> vmOffersTable = vmOffers.getVmOffers();
+		Double cost=0.0;
 		
-		for (Entry<Vm, Integer> entry: vmOffersTable.entrySet()){
+		for (Entry<Vm, Double> entry: vmOffersTable.entrySet()){
 			Vm v = entry.getKey();
 			//use memory due to precision of equality
 			if (v.getRam()==vm.getRam()){
@@ -292,7 +292,7 @@ public class MapReduceDatacenter extends Datacenter {
 	}
 		
 	//Accounts utilization of VMs inside the data center
-	private long updateVmUsage(long startTime, long price) {
+	private long updateVmUsage(long startTime, Double price) {
 		long currentTime = (long) Math.ceil(CloudSim.clock());
 		long runTimeInSeconds = currentTime-startTime;
