+++ cloudsimex-mapreduce/src/org/cloudbus/cloudsim/mapreduce/OffersComparator.java
@@ -9,10 +9,10 @@ import org.cloudbus.cloudsim.Vm;
  * This class implements a comparator for different VM Offers.
  *
  */
-public class OffersComparator implements Comparator<Entry<Vm, Integer>>{
+public class OffersComparator implements Comparator<Entry<Vm, Double>>{
 
 	@Override
-	public int compare(Entry<Vm, Integer> o1,Entry<Vm, Integer> o2) {
+	public int compare(Entry<Vm, Double> o1,Entry<Vm, Double> o2) {
 		
 		if (o1.getValue()>o2.getValue()) return 1;
 		if (o1.getValue()<o2.getValue()) return -1;
